Websocket_Server/requirements.txt
uvicorn==0.30.1
fastapi==0.111.1
paho-mqtt==2.1.0
python-dotenv==1.0.1
logger==1.4